충전하기 누르면 로딩 효과 보여주면서 superpay repository에 선택한 카드, 입력금액 동기화시키고, 끝난 후에 화면 닫기

+++ MiniSuperApp/FinanceHome/Repository/SuperPayRepository.swift
@@ -11,7 +11,7 @@ import Combine
 protocol SuperPayRepository {
   var blance: ReadOnlyCurrentValuePublisher<Double> { get }
   
-  func topup(amount: Int, paymentId: String) -> AnyPublisher<Void, Error>
+  func topup(amount: Double, paymentId: String) -> AnyPublisher<Void, Error>
 }
 
 class SuperPayRepositoryImp: SuperPayRepository {
@@ -20,7 +20,7 @@ class SuperPayRepositoryImp: SuperPayRepository {
   // 원래였으면 서버에서 받아와야 하지만, 이 예제 프로젝트에서는 초기값을 준다.
   private let blanceSubject = CurrentValuePublisher<Double>(0.0)
   
-  func topup(amount: Int, paymentId: String) -> AnyPublisher<Void, Error> {
+  func topup(amount: Double, paymentId: String) -> AnyPublisher<Void, Error> {
     // 서버가 없어서 보이기에 call back 효과를 주기위한 코드
     return Future<Void, Error> { [weak self] promise in
       self?.bgQueue.async {
